[DEV] scheduled_sale,product_custom, sale_order_lien. Herencias y cambios y viustas. Filtro de artículos en progrmaacion

+++ project-addons/scheduled_sale/models/product_template.py
@@ -4,14 +4,15 @@
 
 
 from odoo import _, api, fields, models
-from odoo.exceptions import UserError
-from odoo.osv import expression
+from odoo.exceptions import UserError, ValidationError
+
 
 class ProductTemplate(models.Model):
 
     _inherit = 'product.template'
 
     scheduled_sale_id = fields.Many2one('scheduled.sale', 'Schedule order')
+    origin_scheduled_sale_id = fields.Many2one('scheduled.sale', 'Schedule order')
 
     @api.multi
     def write(self, vals):
@@ -22,79 +23,62 @@ class ProductTemplate(models.Model):
 
 
     @api.multi
-    def unlink_scheduled_templates(self):
+    def archive_scheduled_template(self, scheduled_sale_id = False, undo=True):
+        if not scheduled_sale_id:
+            return
+
         for template in self:
-            template.active = any(x.active for x in x.product_variant_ids)
-        self.filtered(lambda x: not x.active).unlink()
+            if not undo:
+                archive_vals = {'scheduled_sale_id': False, 'origin_scheduled_sale_id': scheduled_sale_id}
+            else:
+                archive_vals = {'scheduled_sale_id': scheduled_sale_id, 'origin_scheduled_sale_id': False}
+            template.write(archive_vals)
 
 
 class ProductProduct(models.Model):
 
     _inherit = 'product.product'
 
-    scheduled_sale_id = fields.Many2one('scheduled.sale', 'Schedule order')
-    origin_scheduled_sale_id = fields.Many2one('scheduled.sale', 'Schedule order')
+    #scheduled_sale_id = fields.Many2one('scheduled.sale', 'Schedule order')
 
     @api.multi
     def action_unlink_product(self):
-        print (self._context)
+
         scheduled_sale_id = self._context.get('scheduled_sale_id', False)
         if not scheduled_sale_id:
             return
-        return self.env['scheduled.sale'].browse(scheduled_sale_id).open_product_to_cancel(self.ids)
+        return self.env['scheduled.sale'].browse(scheduled_sale_id).open_product_to_cancel(self.ids, re_order=self._context.get('re_order', False))
 
 
     @api.multi
     def unlink_scheduled_products(self, scheduled_sale_id=[]):
-        import ipdb; ipdb.set_trace()
-        if not self:
-            return
-        if scheduled_sale_id:
-            scheduled_sale_ids = self.env['scheduled.sale'].browse(scheduled_sale_id)
-        else:
-            scheduled_sale_ids = self.mapped('scheduled_sale_id')
 
-        for scheduled_sale in scheduled_sale_ids:
-            product_ids = self.filtered(lambda x:x.scheduled_sale_id == scheduled_sale)
-            template_ids = product_ids.mapped('product_tmpl_ids')
-            unlink_vals = {'scheduled_sale_id': False,
-                           'origin_scheduled_sale_id': scheduled_sale.id,
-                           'active': False}
+        ##Unlink products if in sale orders
+        ## Implica desactivar los productos, sacarlos de las ordenes de venta y post en el saleorder
 
-            ##sale_orders
-            ## borro todas las lineas de las lineas de venta, donde estén estos articulos y además escribo un mensaje en los pedidos de venta
+        if not self or not scheduled_sale_id:
+            raise ValidationError (_('No products to unlink, or not schedule sale'))
+        ##ventas de estos productos.
+        sale_orders = self.env['sale.order.line'].search([('product_id', 'in', self.ids), ('order_id.scheduled_sale_id','=', scheduled_sale_id), ('state','!=', 'sale')]).mapped('order_id')
 
-            ##ventas de estos productos
-            sale_orders = self.env['sale.order.line'].search([('product_id', 'in', product_ids.ids), ('state','!=', 'sale')]).mapped('order_id')
+        for sale_order in sale_orders:
+            #lineas de venta de estos productos
+            lines = sale_order.mapped('order_line').filtered(lambda x:x.product_id in self)
 
-
-            for sale_order in sale_orders:
-                #lineas de venta de estos productos
-                lines = sale_order.mapped('order_line').filtered(lambda x:x.product_id in product_ids)
-
-                #genero mensaje para sale_order
-                message = _(
+            #genero mensaje para sale_order
+            message = _(
                 "This sale order has been modified and next lines are unlinked from the schedule sale: <a href=# data-oe-model=schedule.sale data-oe-id=%d>%s</a> <ul>") % (
-                      scheduled_sale.id, scheduled_sale.name)
-                for line in lines:
-                    message = _("{} {}".format(message, "<li>{} Qty: {} </li>".format(line.name, line.product_uom_qty)))
-                message = _("{} {}".format(message, "</ul> Date: {}".format(fields.datetime.now())))
-                sale_order. message_post(body=message)
+                scheduled_sale_id, sale_order.scheduled_sale_id.name)
+            for line in lines:
+                message = _("{} {}".format(message, "<li>{} Qty: {} </li>".format(line.name, line.product_uom_qty)))
 
-                lines.unlink()
-            product_ids.write(unlink_vals)
-            template_ids.unlink_scheduled_templates()
+            message = _("{} {}".format(message, "</ul> Date: {}".format(fields.datetime.now())))
+            sale_order.message_post(body=message)
+
+            lines.unlink()
+
+        self.write({'active': False})
         return True
 
-    @api.multi
-    def archive_scheduled_products(self, scheduled_sale_id=False, undo=True):
-        if not scheduled_sale_id:
-            return
-        template_ids = self.mapped('product_tmpl_id')
-        if not undo:
-            archive_vals = {'scheduled_sale_id': False, 'origin_scheduled_sale_id': scheduled_sale_id, 'active': True}
-        else:
-            archive_vals = {'scheduled_sale_id': scheduled_sale_id, 'origin_scheduled_sale_id': False, 'active': True}
-        res = self.write(archive_vals)
-        template_ids.unlink_scheduled_templates()
-        return res
+
+
